% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_vocabs.R
\name{compare_vocabs}
\alias{compare_vocabs}
\title{Compare vocabularies}
\usage{
compare_vocabs(
  tc_dfs,
  word_col = "Word",
  counts_col = "Count",
  weighting_fn = function(x) {     log(x + 1) },
  rel_tf_nrtf_beta = 1,
  zero_negatives = TRUE
)
}
\arguments{
\item{tc_dfs}{Named list of Term Counts data frames.

Each data frame must contain a column with the words and a column with their counts.

Each document must be uniquely named.}

\item{weighting_fn}{Function for weighting the count column before normalizing to frequency.
The metrics using the weighting function will be named with a "_weighted" suffix.

NOTE: This is experimental. The metrics have not been thought through and some may not be meaningful.}
}
\description{
Compare vocabularies
}
\keyword{internal}
