% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_doc_stats.R
\name{get_doc_metrics}
\alias{get_doc_metrics}
\title{Get the metrics for one of the documents}
\usage{
get_doc_metrics(word_scores, doc, remove_zero_counts = TRUE)
}
\arguments{
\item{word_scores}{Scores tibble from \code{compare_vocabs()}.}

\item{doc}{Name of document to retrieve metrics for.}

\item{remove_zero_counts}{Whether to filter out words
that were not in the document. (Logical)}
}
\description{
Get the metrics for one of the documents
}
\author{
Ludvig Renbo Olsen, \email{r-pkgs@ludvigolsen.dk}
}
