% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{hamlet}
\alias{hamlet}
\title{Hamlet lines}
\format{A data frame with 550 rows and 2 variables:
\describe{
\item{Line}{Line said by the character}
\item{Character}{Name of character saying the line}
}}
\source{
\url{https://www.opensourceshakespeare.org/}
}
\description{
Lines by 5 characters from the Hamlet play by William Shakespeare.
}
\author{
Ludvig Renbo Olsen, \email{r-pkgs@ludvigolsen.dk}
}
\keyword{data}
