% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_doc_metrics.R
\name{stack_doc_metrics}
\alias{stack_doc_metrics}
\title{Get the metrics for all documents}
\usage{
stack_doc_metrics(word_scores, remove_zero_counts = TRUE)
}
\arguments{
\item{word_scores}{Scores tibble from \code{compare_vocabs()}.}

\item{remove_zero_counts}{Whether to filter out the words
that were not in a document from that document's results. (Logical)

Often reduces the size of the output but may hinder some downstream analyses.}
}
\description{
Extracts the nested metrics for each document and concatenates them.
}
\author{
Ludvig Renbo Olsen, \email{r-pkgs@ludvigolsen.dk}
}
